// src/interfaces.ts
  export interface Native {
      nld: Nld;
      pap: Pap;
      prs: Prs;
      pus: Pus;
      tuk: Tuk;
      por: Por;
      eng: Eng;
      swe: Swe;
      sqi: Sqi;
      cat: Cat;
      ara: Ara;
      grn: Grn;
      spa: Spa;
      hye: Hye;
      smo: Smo;
      fra: Fra;
      bar: Bar;
      aze: Aze;
      rus: Rus;
      run: Run;
      deu: Deu;
      ben: Ben;
      bul: Bul;
      bos: Bos;
      hrv: Hrv;
      srp: Srp;
      bel: Bel;
      bjz: Bjz;
      aym: Aym;
      que: Que;
      msa: Msa;
      dzo: Dzo;
      nor: Nor;
      tsn: Tsn;
      sag: Sag;
      gsw: Gsw;
      ita: Ita;
      roh: Roh;
      zho: Zho;
      kon: Kon;
      lin: Lin;
      lua: Lua;
      swa: Swa;
      rar: Rar;
      zdj: Zdj;
      ell: Ell;
      tur: Tur;
      ces: Ces;
      slk: Slk;
      dan: Dan;
      tir: Tir;
      ber: Ber;
      mey: Mey;
      est: Est;
      amh: Amh;
      fin: Fin;
      fij: Fij;
      hif: Hif;
      fao: Fao;
      kat: Kat;
      nfr: Nfr;
      pov: Pov;
      kal: Kal;
      cha: Cha;
      hat: Hat;
      hun: Hun;
      ind: Ind;
      glv: Glv;
      hin: Hin;
      tam: Tam;
      gle: Gle;
      fas: Fas;
      arc: Arc;
      ckb: Ckb;
      isl: Isl;
      heb: Heb;
      jam: Jam;
      nrf: Nrf;
      jpn: Jpn;
      kaz: Kaz;
      kir: Kir;
      khm: Khm;
      gil: Gil;
      kor: Kor;
      lao: Lao;
      sin: Sin;
      sot: Sot;
      lit: Lit;
      ltz: Ltz;
      lav: Lav;
      ron: Ron;
      mlg: Mlg;
      div: Div;
      mah: Mah;
      mkd: Mkd;
      mlt: Mlt;
      mya: Mya;
      cnr: Cnr;
      mon: Mon;
      cal: Cal;
      mfe: Mfe;
      nya: Nya;
      afr: Afr;
      her: Her;
      hgm: Hgm;
      kwn: Kwn;
      loz: Loz;
      ndo: Ndo;
      pih: Pih;
      niu: Niu;
      nno: Nno;
      nob: Nob;
      smi: Smi;
      nep: Nep;
      nau: Nau;
      mri: Mri;
      nzs: Nzs;
      urd: Urd;
      fil: Fil;
      pau: Pau;
      hmo: Hmo;
      tpi: Tpi;
      pol: Pol;
      kin: Kin;
      som: Som;
      slv: Slv;
      ssw: Ssw;
      crs: Crs;
      tha: Tha;
      tgk: Tgk;
      tkl: Tkl;
      tet: Tet;
      ton: Ton;
      tvl: Tvl;
      ukr: Ukr;
      uzb: Uzb;
      lat: Lat;
      vie: Vie;
      bis: Bis;
      nbl: Nbl;
      nso: Nso;
      tso: Tso;
      ven: Ven;
      xho: Xho;
      zul: Zul;
      bwg: Bwg;
      kck: Kck;
      khi: Khi;
      ndc: Ndc;
      nde: Nde;
      sna: Sna;
      toi: Toi;
      zib: Zib;
  }

  export interface Name {
      common: string;
      official: string;
      native: Native;
  }

  export interface AWG {
      name: string;
      symbol: string;
  }

  export interface AFN {
      name: string;
      symbol: string;
  }

  export interface AOA {
      name: string;
      symbol: string;
  }

  export interface XCD {
      name: string;
      symbol: string;
  }

  export interface EUR {
      name: string;
      symbol: string;
  }

  export interface ALL {
      name: string;
      symbol: string;
  }

  export interface AED {
      name: string;
      symbol: string;
  }

  export interface ARS {
      name: string;
      symbol: string;
  }

  export interface AMD {
      name: string;
      symbol: string;
  }

  export interface USD {
      name: string;
      symbol: string;
  }

  export interface AUD {
      name: string;
      symbol: string;
  }

  export interface AZN {
      name: string;
      symbol: string;
  }

  export interface BIF {
      name: string;
      symbol: string;
  }

  export interface XOF {
      name: string;
      symbol: string;
  }

  export interface BDT {
      name: string;
      symbol: string;
  }

  export interface BGN {
      name: string;
      symbol: string;
  }

  export interface BHD {
      name: string;
      symbol: string;
  }

  export interface BSD {
      name: string;
      symbol: string;
  }

  export interface BAM {
      name: string;
      symbol: string;
  }

  export interface GBP {
      name: string;
      symbol: string;
  }

  export interface SHP {
      name: string;
      symbol: string;
  }

  export interface BYN {
      name: string;
      symbol: string;
  }

  export interface BZD {
      name: string;
      symbol: string;
  }

  export interface BMD {
      name: string;
      symbol: string;
  }

  export interface BOB {
      name: string;
      symbol: string;
  }

  export interface BRL {
      name: string;
      symbol: string;
  }

  export interface BBD {
      name: string;
      symbol: string;
  }

  export interface BND {
      name: string;
      symbol: string;
  }

  export interface SGD {
      name: string;
      symbol: string;
  }

  export interface BTN {
      name: string;
      symbol: string;
  }

  export interface INR {
      name: string;
      symbol: string;
  }

  export interface BWP {
      name: string;
      symbol: string;
  }

  export interface XAF {
      name: string;
      symbol: string;
  }

  export interface CAD {
      name: string;
      symbol: string;
  }

  export interface CHF {
      name: string;
      symbol: string;
  }

  export interface CLP {
      name: string;
      symbol: string;
  }

  export interface CNY {
      name: string;
      symbol: string;
  }

  export interface CDF {
      name: string;
      symbol: string;
  }

  export interface CKD {
      name: string;
      symbol: string;
  }
// ...
  export interface Currencies {
      AWG: AWG;
      AFN: AFN;
      AOA: AOA;
      XCD: XCD;
      EUR: EUR;
      ALL: ALL;
      AED: AED;
      ARS: ARS;
      AMD: AMD;
      USD: USD;
      AUD: AUD;
      AZN: AZN;
      BIF: BIF;
      XOF: XOF;
      BDT: BDT;
      BGN: BGN;
      BHD: BHD;
      BSD: BSD;
      BAM: BAM;
      GBP: GBP;
      SHP: SHP;
      BYN: BYN;
      BZD: BZD;
      BMD: BMD;
      BOB: BOB;
      BRL: BRL;
      BBD: BBD;
      BND: BND;
      SGD: SGD;
      BTN: BTN;
      INR: INR;
      BWP: BWP;
      XAF: XAF;
      CAD: CAD;
      CHF: CHF;
      CLP: CLP;
      CNY: CNY;
      CDF: CDF;
      CKD: CKD;
      NZD: NZD;
      COP: COP;
      KMF: KMF;
      CVE: CVE;
      CRC: CRC;
      CUC: CUC;
      CUP: CUP;
      ANG: ANG;
      KYD: KYD;
      CZK: CZK;
      DJF: DJF;
      DKK: DKK;
      DOP: DOP;
      DZD: DZD;
      EGP: EGP;
      ERN: ERN;
      MAD: MAD;
      MRU: MRU;
      ETB: ETB;
      FJD: FJD;
      FKP: FKP;
      FOK: FOK;
      GEL: GEL;
      GGP: GGP;
      GHS: GHS;
      GIP: GIP;
      GNF: GNF;
      GMD: GMD;
      GTQ: GTQ;
      GYD: GYD;
      HKD: HKD;
      HNL: HNL;
      HRK: HRK;
      HTG: HTG;
      HUF: HUF;
      IDR: IDR;
      IMP: IMP;
      IRR: IRR;
      IQD: IQD;
      ISK: ISK;
      ILS: ILS;
      JMD: JMD;
      JEP: JEP;
      JOD: JOD;
      JPY: JPY;
      KZT: KZT;
      KES: KES;
      KGS: KGS;
      KHR: KHR;
      KID: KID;
      KRW: KRW;
      KWD: KWD;
      LAK: LAK;
      LBP: LBP;
      LRD: LRD;
      LYD: LYD;
      LKR: LKR;
      LSL: LSL;
      ZAR: ZAR;
      MOP: MOP;
      MDL: MDL;
      MGA: MGA;
      MVR: MVR;
      MXN: MXN;
      MKD: MKD;
      MMK: MMK;
      MNT: MNT;
      MZN: MZN;
      MUR: MUR;
      MWK: MWK;
      MYR: MYR;
      NAD: NAD;
      XPF: XPF;
      NGN: NGN;
      NIO: NIO;
      NOK: NOK;
      NPR: NPR;
      OMR: OMR;
      PKR: PKR;
      PAB: PAB;
      PEN: PEN;
      PHP: PHP;
      PGK: PGK;
      PLN: PLN;
      KPW: KPW;
      PYG: PYG;
      QAR: QAR;
      RON: RON;
      RUB: RUB;
      RWF: RWF;
      SAR: SAR;
      SDG: SDG;
      SBD: SBD;
      SLL: SLL;
      SOS: SOS;
      RSD: RSD;
      SSP: SSP;
      STN: STN;
      SRD: SRD;
      SEK: SEK;
      SZL: SZL;
      SCR: SCR;
      SYP: SYP;
      THB: THB;
      TJS: TJS;
      TMT: TMT;
      TOP: TOP;
      TTD: TTD;
      TND: TND;
      TRY: TRY;
      TVD: TVD;
      TWD: TWD;
      TZS: TZS;
      UGX: UGX;
      UAH: UAH;
      UYU: UYU;
      UZS: UZS;
      VES: VES;
      VND: VND;
      VUV: VUV;
      WST: WST;
      YER: YER;
      ZMW: ZMW;
      ZWB: ZWB;
  }

  export interface Idd {
      root: string;
      suffixes: string[];
  }

  export interface Languages {
      nld: string;
      pap: string;
      prs: string;
      pus: string;
      tuk: string;
      por: string;
      eng: string;
      swe: string;
      sqi: string;
      cat: string;
      ara: string;
      grn: string;
      spa: string;
      hye: string;
      smo: string;
      fra: string;
      bar: string;
      aze: string;
      rus: string;
      run: string;
      deu: string;
      ben: string;
      bul: string;
      bos: string;
      hrv: string;
      srp: string;
      bel: string;
      bjz: string;
      aym: string;
      que: string;
      msa: string;
      dzo: string;
      nor: string;
      tsn: string;
      sag: string;
      gsw: string;
      ita: string;
      roh: string;
      zho: string;
      kon: string;
      lin: string;
      lua: string;
      swa: string;
      rar: string;
      zdj: string;
      ell: string;
      tur: string;
      ces: string;
      slk: string;
      dan: string;
      tir: string;
      ber: string;
      mey: string;
      est: string;
      amh: string;
      fin: string;
      fij: string;
      hif: string;
      fao: string;
      kat: string;
      nfr: string;
      pov: string;
      kal: string;
      cha: string;
      hat: string;
      hun: string;
      ind: string;
      glv: string;
      hin: string;
      tam: string;
      gle: string;
      fas: string;
      arc: string;
      ckb: string;
      isl: string;
      heb: string;
      jam: string;
      nrf: string;
      jpn: string;
      kaz: string;
      kir: string;
      khm: string;
      gil: string;
      kor: string;
      lao: string;
      sin: string;
      sot: string;
      lit: string;
      ltz: string;
      lav: string;
      ron: string;
      mlg: string;
      div: string;
      mah: string;
      mkd: string;
      mlt: string;
      mya: string;
      cnr: string;
      mon: string;
      cal: string;
      mfe: string;
      nya: string;
      afr: string;
      her: string;
      hgm: string;
      kwn: string;
      loz: string;
      ndo: string;
      pih: string;
      niu: string;
      nno: string;
      nob: string;
      smi: string;
      nep: string;
      nau: string;
      mri: string;
      nzs: string;
      urd: string;
      fil: string;
      pau: string;
      hmo: string;
      tpi: string;
      pol: string;
      kin: string;
      som: string;
      slv: string;
      ssw: string;
      crs: string;
      tha: string;
      tgk: string;
      tkl: string;
      tet: string;
      ton: string;
      tvl: string;
      ukr: string;
      uzb: string;
      lat: string;
      vie: string;
      bis: string;
      nbl: string;
      nso: string;
      tso: string;
      ven: string;
      xho: string;
      zul: string;
      bwg: string;
      kck: string;
      khi: string;
      ndc: string;
      nde: string;
      sna: string;
      toi: string;
      zib: string;
  }
// ...
  export interface Translations {
      ces: Ces2;
      deu: Deu2;
      est: Est2;
      fin: Fin2;
      fra: Fra2;
      hrv: Hrv2;
      hun: Hun2;
      ita: Ita2;
      jpn: Jpn2;
      kor: Kor2;
      nld: Nld2;
      per: Per;
      pol: Pol2;
      por: Por2;
      rus: Rus2;
      slk: Slk2;
      spa: Spa2;
      swe: Swe2;
      urd: Urd2;
      zho: Zho2;
      cym: Cym;
  }

  export interface Eng2 {
      f: string;
      m: string;
  }

  export interface Fra3 {
      f: string;
      m: string;
  }

  export interface Demonyms {
      eng: Eng2;
      fra: Fra3;
  }

  export interface RootObject {
      name: Name;
      tld: string[];
      alpha2Code: string;
      alpha3Code: string;
      independent?: boolean;
      status: string;
      unMember: boolean;
      currencies: Currencies;
      idd: Idd;
      capital: string[];
      altSpellings: string[];
      region: string;
      subregion: string;
      languages: Languages;
      translations: Translations;
      latlng: number[];
      landlocked: boolean;
      borders: string[];
      area: number;
      demonyms: Demonyms;
  }